refactor + new test
