+++ app/api/entities/entitySavingManager.ts
@@ -1,22 +1,9 @@
-/* eslint-disable max-lines */
-import { WithId } from 'api/odm';
-import {
-  attachmentsPath,
-  uploadsPath,
-  files as filesAPI,
-  generateFileName,
-  storeFile,
-} from 'api/files';
-import { search } from 'api/search';
-import { errorLog } from 'api/log';
-import entities from 'api/entities/entities';
-import { prettifyError } from 'api/utils/handleError';
-import { EntityWithFilesSchema } from 'shared/types/entityType';
-import { FileType } from 'shared/types/fileType';
-import { UserSchema } from 'shared/types/userType';
-import { MetadataObjectSchema } from 'shared/types/commonTypes';
-import { processDocument } from 'api/files/processDocument';
 import { set } from 'lodash';
+import { generateFileName } from 'api/files';
+import entities from 'api/entities/entities';
+import { EntityWithFilesSchema } from 'shared/types/entityType';
+import { UserSchema } from 'shared/types/userType';
+import { handleAttachmentInMetadataProperties, processFiles, saveFiles } from './managerFunctions';
 
 type FileAttachments = {
   originalname: string;
@@ -27,168 +14,6 @@ type FileAttachments = {
   encoding?: string;
 };
 
-async function prepareNewFiles(
-  entity: EntityWithFilesSchema,
-  updatedEntity: EntityWithFilesSchema,
-  fileAttachments: FileType[] = [],
-  documentAttachments: FileType[] = []
-) {
-  const attachments: FileType[] = [];
-  const documents: FileType[] = [];
-  const newUrls = entity.attachments?.filter(attachment => !attachment._id && attachment.url);
-
-  if (fileAttachments.length) {
-    await Promise.all(
-      fileAttachments.map(async (file: FileType) => {
-        const savedFile = await storeFile(attachmentsPath, file, true);
-        attachments.push({
-          ...savedFile,
-          entity: updatedEntity.sharedId,
-          type: 'attachment',
-        });
-      })
-    );
-  }
-
-  if (newUrls && newUrls.length) {
-    newUrls.map(async (url: any) => {
-      attachments.push({
-        ...url,
-        entity: updatedEntity.sharedId,
-        type: 'attachment',
-      });
-    });
-  }
-
-  if (documentAttachments.length) {
-    await Promise.all(
-      documentAttachments.map(async (document: FileType) => {
-        const savedDocument = await storeFile(uploadsPath, document, true);
-        documents.push({
-          ...savedDocument,
-          entity: updatedEntity.sharedId,
-          type: 'document',
-        });
-      })
-    );
-  }
-
-  return { attachments, documents };
-}
-
-const updateDeletedFiles = async (
-  entityFiles: WithId<FileType>[],
-  entity: EntityWithFilesSchema,
-  type: 'attachment' | 'document'
-) => {
-  const deletedFiles = entityFiles.filter(
-    existingFile =>
-      existingFile._id &&
-      existingFile.type === type &&
-      !entity[type === 'attachment' ? 'attachments' : 'documents']?.find(
-        attachment => attachment._id?.toString() === existingFile._id.toString()
-      )
-  );
-  await Promise.all(deletedFiles.map(async file => filesAPI.delete(file)));
-};
-
-const filterRenamedFiles = (entity: EntityWithFilesSchema, entityFiles: WithId<FileType>[]) => {
-  const renamedAttachments = entity.attachments
-    ? entity.attachments
-        .filter(
-          (attachment: FileType) =>
-            attachment._id &&
-            entityFiles.find(
-              file =>
-                attachment._id?.toString() === file._id.toString() &&
-                file.originalname !== attachment.originalname
-            )
-        )
-        .map((attachment: FileType) => ({
-          _id: attachment._id!.toString(),
-          originalname: attachment.originalname,
-        }))
-    : [];
-
-  const renamedDocuments = entity.documents
-    ? entity.documents
-        .filter(
-          (document: FileType) =>
-            document._id &&
-            entityFiles.find(
-              file =>
-                document._id?.toString() === file._id.toString() &&
-                file.originalname !== document.originalname
-            )
-        )
-        .map((document: FileType) => ({
-          _id: document._id!.toString(),
-          originalname: document.originalname,
-        }))
-    : [];
-
-  return { renamedAttachments, renamedDocuments };
-};
-
-const processFiles = async (
-  entity: EntityWithFilesSchema,
-  updatedEntity: EntityWithFilesSchema,
-  fileAttachments: FileType[] | undefined,
-  documentAttachments: FileType[] | undefined
-) => {
-  const { attachments, documents } = await prepareNewFiles(
-    entity,
-    updatedEntity,
-    fileAttachments,
-    documentAttachments
-  );
-
-  if (entity._id && (entity.attachments || entity.documents)) {
-    const entityFiles: WithId<FileType>[] = await filesAPI.get(
-      { entity: entity.sharedId, type: { $in: ['attachment', 'document'] } },
-      '_id, originalname, type'
-    );
-
-    await updateDeletedFiles(entityFiles, entity, 'attachment');
-    await updateDeletedFiles(entityFiles, entity, 'document');
-
-    const { renamedAttachments, renamedDocuments } = filterRenamedFiles(entity, entityFiles);
-
-    attachments.push(...renamedAttachments);
-    documents.push(...renamedDocuments);
-  }
-
-  return { proccessedAttachments: attachments, proccessedDocuments: documents };
-};
-
-const bindAttachmentToMetadataProperty = (
-  _values: MetadataObjectSchema[],
-  attachments: FileType[]
-) => {
-  const values = _values;
-  if (_values[0].attachment !== undefined) {
-    values[0].value = attachments[_values[0].attachment]
-      ? `/api/files/${attachments[_values[0].attachment].filename}`
-      : '';
-  }
-  return values;
-};
-
-const handleAttachmentInMetadataProperties = (
-  entity: EntityWithFilesSchema,
-  attachments: FileType[]
-) => {
-  Object.entries(entity.metadata || {}).forEach(([_property, _values]) => {
-    if (_values && _values.length) {
-      const values = bindAttachmentToMetadataProperty(_values, attachments);
-      delete values[0].attachment;
-    }
-  });
-
-  return entity;
-};
-
-// eslint-disable-next-line max-statements
 const saveEntity = async (
   _entity: EntityWithFilesSchema,
   {
@@ -197,8 +22,6 @@ const saveEntity = async (
     files: reqFiles,
   }: { user: UserSchema; language: string; files?: FileAttachments[] }
 ) => {
-  const fileSaveErrors: string[] = [];
-
   const { attachments, documents } = (reqFiles || []).reduce(
     (acum, file) => {
       const namedFile = { ...file, filename: generateFileName(file) };
@@ -225,32 +48,7 @@ const saveEntity = async (
     documents
   );
 
-  if (proccessedAttachments.length) {
-    await Promise.all(
-      proccessedAttachments.map(async attachment => {
-        try {
-          await filesAPI.save(attachment, false);
-        } catch (e) {
-          errorLog.error(prettifyError(e));
-          fileSaveErrors.push(`Could not save supporting file/s: ${attachment.originalname}`);
-        }
-      })
-    );
-    await search.indexEntities({ sharedId: updatedEntity.sharedId }, '+fullText');
-  }
-
-  if (proccessedDocuments.length) {
-    await Promise.all(
-      proccessedDocuments.map(async document => {
-        try {
-          await processDocument(updatedEntity.sharedId, document);
-        } catch (e) {
-          errorLog.error(prettifyError(e));
-          fileSaveErrors.push(`Could not save pdf file/s: ${document.originalname}`);
-        }
-      })
-    );
-  }
+  const fileSaveErrors = await saveFiles(proccessedAttachments, proccessedDocuments, updatedEntity);
 
   const [entityWithAttachments] = await entities.getUnrestrictedWithDocuments(
     {
